requirements.txt
# Core GUI Framework
PyQt5>=5.15.0

# Data Processing
pandas>=1.3.0
openpyxl>=3.0.0
numpy>=1.21.0

# Screen Automation
pyautogui>=0.9.53
pillow>=8.3.0
opencv-python>=4.5.0
pynput>=1.7.0

# Screen Information
screeninfo>=0.8.0

# Image Processing
mss>=6.1.0

# OCR (Optical Character Recognition)
paddlepaddle>=2.5.0  # PaddlePaddle framework (CPU version)
# paddlepaddle-gpu>=2.5.0  # Uncomment for GPU version
paddleocr>=2.7.0  # PaddleOCR

# Security
cryptography>=3.4.0

# Development tools (optional)
black>=21.0
flake8>=3.9.0